+++ app/db/Vertex.js
