+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "simian-store",
-  "version": "0.0.1",
+  "version": "1.0.0",
   "description": "CRUD for the simian photo server project.",
-  "main": "lib/index.js",
+  "main": "app/index.js",
   "scripts": {
     "document": "jsdoc -c jsdoc.conf",
     "fix-lint": "eslint lib/ --fix",
@@ -24,7 +24,7 @@
   "nyc": {
     "all": true,
     "include": [
-      "lib/*.js"
+      "app/"
     ],
     "exclude": [
       "node_modules",
